Replaced BlockEvent for custom packets, better SMP support
I was using the blockparam from addBlockEvent to send the current slot
and the record Id, but even though the variable is integer, inside the
S24PacketBlockAction packet it gets converted to byte, so I can't use as
I intended.

Custom packets is more reliable and also SMP friendly.

Also fixed the BreakBlock moethod that was causing crashes in SMP.